# Source code/swaggerAPI.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Human Resource Management System",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/hrms",
      "variables": {}
    }
  ],
  "paths": {
...
    "/user/manage": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Manage User",
        "operationId": "ManageUser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManageUserRequest"
              },
              "example": {
                "subordinateId": 3,
                "managerId": 4
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/user/infor/uploadAvatar": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Upload Avatar",
        "operationId": "UploadAvatar",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "encoding": {},
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ManageUserRequest": {
        "title": "ManageUserRequest",
        "required": [
          "subordinateId",
          "managerId"
        ],
        "type": "object",
        "properties": {
          "subordinateId": {
            "type": "integer",
            "format": "int32"
          },
          "managerId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "subordinateId": 3,
          "managerId": 4
        }
      },
...
    },
    "securitySchemes": {
      "httpBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "httpBearer": []
    }
  ],
...
}